fix tests

+++ engine/src/flutter/shell/platform/android/test/io/flutter/view/AccessibilityBridgeTest.java
@@ -2097,7 +2097,7 @@ public class AccessibilityBridgeTest {
     // These fields are declared in the order they should be
     // encoded.
     int id = 0;
-    int flags = 0;
+    long flags = 0;
     int actions = 0;
     int maxValueLength = 0;
     int currentValueLength = 0;
@@ -2156,7 +2156,7 @@ public class AccessibilityBridgeTest {
     protected void addToBuffer(
         ByteBuffer bytes, ArrayList<String> strings, ArrayList<ByteBuffer> stringAttributeArgs) {
       bytes.putInt(id);
-      bytes.putInt(flags);
+      bytes.putLong(flags);
       bytes.putInt(actions);
       bytes.putInt(maxValueLength);
       bytes.putInt(currentValueLength);
